Fix: Test never called asLong as the test case suggested.

+++ src/test/java/io/lordtylus/equation/ConstantTest.java
@@ -112,7 +112,7 @@ class ConstantTest {
 
         /* Then */
 
-        long value = actual.asInt();
+        long value = actual.asLong();
 
         assertEquals(123, value);
     }
