+++ src/test/java/io/lordtylus/equation/VariableTest.java
@@ -125,7 +125,7 @@ class VariableTest {
 
         /* Then */
 
-        long value = actual.asInt();
+        long value = actual.asLong();
 
         assertEquals(123, value);
     }
